feat: add pasword textfield PSA-50
